requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
cycler==0.10.0
discord.py==1.6.0
idna==3.1
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
mpl-interactions==0.14.1
multidict==5.1.0
numpy==1.20.0
pandas==1.2.0
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
yarl==1.6.3
